requirements.txt
# requirements.txt
pandas
requests
snowflake-connector-python
snowflake-snowpark-python
streamlit==1.31.1
watchdog